backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-dotenv
torch
transformers
boto3
mangum 
python-jose[cryptography]
python-whois
pytest
pytest-cov
httpx
python-jwt
PyJWT
bcrypt
requests